requirements.txt
# ТОЧНЫЕ ВЕРСИИ - НЕ ИЗМЕНЯТЬ!
web3==6.11.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
customtkinter==5.2.1
pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
click==8.1.7
pydantic==2.5.0
alembic==1.12.1
tenacity==8.2.3
schedule==1.2.0
